refactor out of main file
