+++ lib/resque_bus/publishing.rb
@@ -1,0 +1,119 @@
+module ResqueBus
+  module Publishing
+    def with_global_attributes(attributes)
+      original_timezone = false
+      original_locale   = false
+      
+      if attributes["bus_locale"] && defined?(I18n) && I18n.respond_to?(:locale=)
+        original_locale = I18n.locale if I18n.respond_to?(:locale)
+        I18n.locale = attributes["bus_locale"]
+      end
+      
+      if attributes["bus_timezone"] && defined?(Time) && Time.respond_to?(:zone=)
+        original_timezone = Time.zone if Time.respond_to?(:zone)
+        Time.zone = attributes["bus_timezone"]
+      end
+
+      yield
+    ensure
+      I18n.locale = original_locale   unless original_locale   == false
+      Time.zone   = original_timezone unless original_timezone == false
+    end
+    
+    def publish_metadata(event_type, attributes={})
+      # TODO: "bus_app_key" => application.app_key ?
+      bus_attr = {"bus_published_at" => Time.now.to_i, "bus_event_type" => event_type}
+      bus_attr["bus_id"]           = "#{Time.now.to_i}-#{generate_uuid}"
+      bus_attr["bus_app_hostname"] = ResqueBus.hostname
+      bus_attr["bus_locale"]       = I18n.locale.to_s if defined?(I18n) && I18n.respond_to?(:locale) && I18n.locale
+      bus_attr["bus_timezone"]     = Time.zone.name   if defined?(Time) && Time.respond_to?(:zone)   && Time.zone
+      out = bus_attr.merge(attributes || {})
+      ResqueBus.before_publish_callback(out)
+      out
+    end
+    
+    def generate_uuid
+      require 'securerandom' unless defined?(SecureRandom)
+      return SecureRandom.uuid
+      
+      rescue Exception => e
+        # secure random not there
+        # big random number a few times
+        n_bytes = [42].pack('i').size
+        n_bits = n_bytes * 8
+        max = 2 ** (n_bits - 2) - 1
+        return "#{rand(max)}-#{rand(max)}-#{rand(max)}"
+    end
+    
+    def publish(event_type, attributes = {})
+      to_publish = publish_metadata(event_type, attributes)
+      ResqueBus.log_application("Event published: #{event_type} #{to_publish.inspect}")
+      if local_mode
+        ResqueBus::Local.perform(to_publish)
+      else
+        enqueue_to(incoming_queue, Driver, to_publish)
+      end
+    end
+    
+    def publish_at(timestamp_or_epoch, event_type, attributes = {})
+      to_publish = publish_metadata(event_type, attributes)
+      to_publish["bus_delayed_until"] ||= timestamp_or_epoch.to_i
+      to_publish.delete("bus_published_at") unless attributes["bus_published_at"] # will be put on when it actually does it
+      
+      ResqueBus.log_application("Event published:#{event_type} #{to_publish.inspect} publish_at: #{timestamp_or_epoch.to_i}")
+      item = delayed_job_to_hash_with_queue(incoming_queue, Publisher, [event_type, to_publish])
+      delayed_push(timestamp_or_epoch, item)
+    end
+
+    def enqueue_to(queue, klass, *args)
+      push(queue, :class => klass.to_s, :args => args)
+    end
+
+    ## From Resque, but using a (possibly) different instance of Redis
+    
+    # Pushes a job onto a queue. Queue name should be a string and the
+    # item should be any JSON-able Ruby object.
+    #
+    # Resque works generally expect the `item` to be a hash with the following
+    # keys:
+    #
+    #   class - The String name of the job to run.
+    #    args - An Array of arguments to pass the job. Usually passed
+    #           via `class.to_class.perform(*args)`.
+    #
+    # Example
+    #
+    #   Resque.push('archive', :class => 'Archive', :args => [ 35, 'tar' ])
+    #
+    # Returns nothing
+    def push(queue, item)
+      watch_queue(queue)
+      ResqueBus.redis.rpush "queue:#{queue}", Resque.encode(item)
+    end
+    
+    # Used internally to keep track of which queues we've created.
+    # Don't call this directly.
+    def watch_queue(queue)
+      ResqueBus.redis.sadd(:queues, queue.to_s)
+    end
+    
+    ### From Resque Scheduler
+    # Used internally to stuff the item into the schedule sorted list.
+    # +timestamp+ can be either in seconds or a datetime object
+    # Insertion if O(log(n)).
+    # Returns true if it's the first job to be scheduled at that time, else false
+    def delayed_push(timestamp, item)
+      # First add this item to the list for this timestamp
+      ResqueBus.redis.rpush("delayed:#{timestamp.to_i}", Resque.encode(item))
+
+      # Now, add this timestamp to the zsets.  The score and the value are
+      # the same since we'll be querying by timestamp, and we don't have
+      # anything else to store.
+      ResqueBus.redis.zadd :delayed_queue_schedule, timestamp.to_i, timestamp.to_i
+    end
+
+    def delayed_job_to_hash_with_queue(queue, klass, args)
+      {:class => klass.to_s, :args => args, :queue => queue}
+    end
+  end
+end
